\name{nozeroprices}
\Rdversion{1.1}
\alias{nozeroprices}
\title{
Delete the observations where the price is zero
}
\description{
Function deletes the observations where the price is zero.
}

\usage{
nozeroprices(tdata);
}

\arguments{
  \item{tdata}{an xts object at least containing a column "PRICE".} 
}

\section{Details}{}

\value{
xts object
}
\references{
}
\author{ Jonathan Cornelissen and Kris Boudt}