\name{rmnegspread}
\Rdversion{1.1}
\alias{rmnegspread}
\title{
Delete entries for which the spread is negative
}
\description{
Function deletes entries for which the spread is negative.
}

\usage{
rmnegspread(qdata);
}

\arguments{
  \item{qdata}{an xts object at least containing the columns "BID" and "OFFER".} 
}

\section{Details}{}

\value{
xts object
}
\references{
}
\author{ Jonathan Cornelissen and Kris Boudt}